refactor(client): Split DebugPanel.tsx into tabbed modules for LLM optimization
- Split 1208-line DebugPanel.tsx into focused modules:
  - DebugPanel.tsx: Main orchestrator with tab navigation (70 lines)
  - GameplayTab.tsx: Combat, cards, mana, units, level, movement (400 lines)
  - SaveLoadTab.tsx: Quick slots, import/export (120 lines)
  - AudioTab.tsx: Sound effects, ambient music (250 lines)
  - debugPanelData.ts: Constants and data definitions (240 lines)
  - types.ts: Shared types (25 lines)
  - index.ts: Clean exports

- Added tabbed interface with Gameplay, Save/Load, and Audio tabs
- Improves LLM context efficiency by allowing focused file reads
- Removes emoji icons for cleaner text-based UI
