+++ packages/client/src/components/DebugPanel/index.ts
@@ -1,0 +1,15 @@
+/**
+ * DebugPanel Module
+ *
+ * Development tools for testing the Mage Knight game.
+ * Split into focused modules for LLM optimization:
+ *
+ * - DebugPanel.tsx: Main orchestrator with tab navigation
+ * - GameplayTab.tsx: Combat, cards, mana, units, level, movement controls
+ * - SaveLoadTab.tsx: Quick slots, import/export
+ * - AudioTab.tsx: Sound effects, ambient music controls
+ * - debugPanelData.ts: Constants and data definitions
+ * - types.ts: Shared types
+ */
+
+export { DebugPanel } from "./DebugPanel";
